[refactor] test 코드 수정
